refactor(plugins): Decouple jackson from core models
